+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/iggy/conversion/SwfShapeToIggyConvertor.java
@@ -37,12 +37,12 @@ public class SwfShapeToIggyConvertor {
         return (val / 1024f);
     }
 
-    public static IggyShape convertShape(SHAPE swfShape) {
+    public static IggyShape convertShape(int shapeNum, SHAPE swfShape) {
         /*if (swfShape.shapeRecords.size() == 1) { //no glyphs, maybe space
             return null;
         }*/
         List<IggyShapeNode> nodes = new ArrayList<>();
-        RECT bounds = swfShape.getBounds();
+        RECT bounds = swfShape.getBounds(shapeNum);
         boolean first = true;
         float curX = 0f;
         float curY = 0f;