+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/types/shaperecords/SHAPERECORD.java
@@ -24,6 +24,9 @@ import com.jpexs.decompiler.flash.helpers.FontHelper;
 import com.jpexs.decompiler.flash.tags.base.NeedsCharacters;
 import com.jpexs.decompiler.flash.tags.base.TextTag;
 import com.jpexs.decompiler.flash.types.ColorTransform;
+import com.jpexs.decompiler.flash.types.LINESTYLE;
+import com.jpexs.decompiler.flash.types.LINESTYLE2;
+import com.jpexs.decompiler.flash.types.LINESTYLEARRAY;
 import com.jpexs.decompiler.flash.types.MATRIX;
 import com.jpexs.decompiler.flash.types.RECT;
 import com.jpexs.decompiler.flash.types.RGB;
@@ -77,7 +80,7 @@ public abstract class SHAPERECORD implements Cloneable, NeedsCharacters, Seriali
 
     public abstract void flip();
 
-    public static RECT getBounds(List<SHAPERECORD> records) {
+    public static RECT getBounds(List<SHAPERECORD> records, LINESTYLEARRAY lineStyles, int shapeNum) {
         int x = 0;
         int y = 0;
         int max_x = 0;
@@ -85,7 +88,29 @@ public abstract class SHAPERECORD implements Cloneable, NeedsCharacters, Seriali
         int min_x = Integer.MAX_VALUE;
         int min_y = Integer.MAX_VALUE;
         boolean started = false;
+        int lineStyle = 0;
+        int lineWidth = 0;
+        int lineWidthHalf = 0;
         for (SHAPERECORD r : records) {
+            if (r instanceof StyleChangeRecord) {
+                StyleChangeRecord style = (StyleChangeRecord) r;
+                if (style.stateNewStyles) {
+                    lineStyles = style.lineStyles;
+                }
+                if (style.stateLineStyle) {
+                    lineStyle = style.lineStyle;
+                    if (lineStyle == 0) {
+                        lineWidth = 0;
+                    } else {
+                        if (shapeNum <= 3) {
+                            lineWidth = lineStyles.lineStyles[lineStyle - 1].width;
+                        } else {
+                            lineWidth = lineStyles.lineStyles2[lineStyle - 1].width;
+                        }
+                    }
+                    lineWidthHalf = lineWidth / 2;
+                }
+            }
             if (r instanceof CurvedEdgeRecord) {
                 CurvedEdgeRecord curverEdge = (CurvedEdgeRecord) r;
                 int x2 = x + curverEdge.controlDeltaX;
@@ -96,12 +121,26 @@ public abstract class SHAPERECORD implements Cloneable, NeedsCharacters, Seriali
                 if (y2 > max_y) {
                     max_y = y2;
                 }
+                if (x2 + lineWidthHalf > max_x) {
+                    max_x = x2 + lineWidthHalf;
+                }
+                if (y2 + lineWidthHalf > max_y) {
+                    max_y = y2 + lineWidthHalf;
+                }
                 if (started) {
                     if (y2 < min_y) {
                         min_y = y2;
                     }
                     if (x2 < min_x) {
                         min_x = x2;
+                    }
+
+                    if (y2 - lineWidthHalf < min_y) {
+                        min_y = y2 - lineWidthHalf;
+                    }
+
+                    if (x2 - lineWidthHalf < min_x) {
+                        min_x = x2 - lineWidthHalf;
                     }
                 }
             }
@@ -114,6 +153,12 @@ public abstract class SHAPERECORD implements Cloneable, NeedsCharacters, Seriali
             if (y > max_y) {
                 max_y = y;
             }
+            if (x + lineWidthHalf > max_x) {
+                max_x = x + lineWidthHalf;
+            }
+            if (y + lineWidthHalf > max_y) {
+                max_y = y + lineWidthHalf;
+            }
             if (r.isMove()) {
                 started = true;
             }
@@ -123,6 +168,13 @@ public abstract class SHAPERECORD implements Cloneable, NeedsCharacters, Seriali
                 }
                 if (x < min_x) {
                     min_x = x;
+                }
+
+                if (y - lineWidthHalf < min_y) {
+                    min_y = y - lineWidthHalf;
+                }
+                if (x - lineWidthHalf < min_x) {
+                    min_x = x - lineWidthHalf;
                 }
             }
         }
@@ -150,8 +202,11 @@ public abstract class SHAPERECORD implements Cloneable, NeedsCharacters, Seriali
         int maxh = 0;
         int minXMin = 0;
         int minYMin = 0;
+        LINESTYLEARRAY lsa = new LINESTYLEARRAY();
+        lsa.lineStyles = new LINESTYLE[0];
+        lsa.lineStyles2 = new LINESTYLE2[0];
         for (SHAPE s : shapes) {
-            RECT r = SHAPERECORD.getBounds(s.shapeRecords);
+            RECT r = SHAPERECORD.getBounds(s.shapeRecords, lsa, shapeNum);
             if (r.Xmax < r.Xmin || r.Ymax < r.Ymin) {
                 continue;
             }
@@ -207,7 +262,7 @@ public abstract class SHAPERECORD implements Cloneable, NeedsCharacters, Seriali
                 // shapeNum: 1
                 SHAPE shape = shapes.get(pos);
                 List<SHAPERECORD> records = shape.shapeRecords;
-                RECT bounds = SHAPERECORD.getBounds(records);
+                RECT bounds = SHAPERECORD.getBounds(records, lsa, shapeNum);
 
                 int w1 = bounds.getWidth();
                 int h1 = bounds.getHeight();